+++ libs/core-ui/src/index.ts
@@ -27,6 +27,7 @@ export * from "./lib/util/FabricStyles";
 export * from "./lib/util/JointDataset";
 export * from "./lib/util/generateRoute";
 export * from "./lib/util/getRandomId";
+export * from "./lib/util/getCohortFilterCount";
 export * from "./lib/util/IGenericChartProps";
 export * from "./lib/util/initializeOfficeFabric";
 export * from "./lib/util/ModelExplanationUtils";
@@ -34,7 +35,10 @@ export * from "./lib/util/Never";
 export * from "./lib/util/PartialRequired";
 export * from "./lib/util/nameof";
 export * from "./lib/util/rowErrorSize";
+export * from "./lib/util/getBasicFilterString";
+export * from "./lib/util/getCompositeFilterString";
 export * from "./lib/util/getFeatureOptions";
+export * from "./lib/util/getFilterBoundsArgs";
 export * from "./lib/util/string";
 export * from "./lib/util/toScientific";
 export * from "./lib/util/ITelemetryMessage";
