+++ libs/model-assessment/src/lib/ModelAssessmentDashboard/Cohort/CreateGlobalCohort.tsx
@@ -38,10 +38,12 @@ export class CreateGlobalCohort extends React.Component<ICreateGlobalCohortProps
       )
     );
   }
+
   private addCohort = (cohort: Cohort, switchNew?: boolean) => {
     this.context.addCohort(cohort, switchNew);
     this.props.onDismiss();
   };
+
   private getExistingCohortName(): string[] {
     return this.context.errorCohorts
       .filter((errorCohort) => !errorCohort.isTemporary)
