+++ libs/model-assessment/src/lib/ModelAssessmentDashboard/Cohort/ShiftCohort.tsx
@@ -72,9 +72,6 @@ export class ShiftCohort extends React.Component<
     }
     const filters =
       this.state.savedCohorts[this.state.selectedCohort].cohort.filters;
-    const compositeFilters =
-      this.state.savedCohorts[this.state.selectedCohort].cohort
-        .compositeFilters;
     return (
       <Dialog
         hidden={false}
@@ -103,8 +100,11 @@ export class ShiftCohort extends React.Component<
           onChange={this.onChange}
         />
         <CohortEditorFilterList
+          compositeFilters={
+            this.state.savedCohorts[this.state.selectedCohort].cohort
+              .compositeFilters
+          }
           filters={filters}
-          compositeFilters={compositeFilters}
           jointDataset={this.context.jointDataset}
         />
         <DialogFooter>
